fix:transfer ownership,notification

+++ priv/repo/migrations/20250212102142_alter_user_references_in_basic_profile.exs
@@ -1,5 +1,6 @@
 defmodule WraftDoc.Repo.Migrations.AlterUserReferencesInBasicProfile do
   use Ecto.Migration
+  alias WraftDoc.Enterprise
 
   def up do
     drop(constraint("basic_profile", "basic_profile_user_id_fkey"))
@@ -10,22 +11,27 @@ defmodule WraftDoc.Repo.Migrations.AlterUserReferencesInBasicProfile do
       )
     end
 
-    drop(constraint("subscriptions", "subscriptions_plan_id_fkey"))
-    drop(constraint("subscriptions", "subscriptions_organisation_id_fkey"))
-    drop(constraint("subscriptions", "subscriptions_subscriber_id_fkey"))
+    unless Enterprise.self_hosted?() do
+      drop(constraint("subscriptions", "subscriptions_plan_id_fkey"))
+      drop(constraint("subscriptions", "subscriptions_organisation_id_fkey"))
+      drop(constraint("subscriptions", "subscriptions_subscriber_id_fkey"))
 
-    alter table(:subscriptions) do
-      modify(:plan_id, references(:plan, type: :uuid, column: :id, on_delete: :nilify_all),
-        null: false
-      )
+      alter table(:subscriptions) do
+        modify(:plan_id, references(:plan, type: :uuid, column: :id, on_delete: :nilify_all),
+          null: false
+        )
 
-      modify(
-        :organisation_id,
-        references(:organisation, type: :uuid, column: :id, on_delete: :delete_all),
-        null: false
-      )
+        modify(
+          :organisation_id,
+          references(:organisation, type: :uuid, column: :id, on_delete: :delete_all),
+          null: false
+        )
 
-      modify(:subscriber_id, references(:user, type: :uuid, column: :id, on_delete: :nilify_all))
+        modify(
+          :subscriber_id,
+          references(:user, type: :uuid, column: :id, on_delete: :nilify_all)
+        )
+      end
     end
   end
 
@@ -38,22 +44,24 @@ defmodule WraftDoc.Repo.Migrations.AlterUserReferencesInBasicProfile do
       )
     end
 
-    drop(constraint("subscriptions", "subscriptions_plan_id_fkey"))
-    drop(constraint("subscriptions", "subscriptions_organisation_id_fkey"))
-    drop(constraint("subscriptions", "subscriptions_subscriber_id_fkey"))
+    unless Enterprise.self_hosted?() do
+      drop(constraint("subscriptions", "subscriptions_plan_id_fkey"))
+      drop(constraint("subscriptions", "subscriptions_organisation_id_fkey"))
+      drop(constraint("subscriptions", "subscriptions_subscriber_id_fkey"))
 
-    alter table(:subscriptions) do
-      modify(:plan_id, references(:plan, type: :uuid, column: :id, on_delete: :nothing),
-        null: false
-      )
+      alter table(:subscriptions) do
+        modify(:plan_id, references(:plan, type: :uuid, column: :id, on_delete: :nothing),
+          null: false
+        )
 
-      modify(
-        :organisation_id,
-        references(:organisation, type: :uuid, column: :id, on_delete: :nothing),
-        null: false
-      )
+        modify(
+          :organisation_id,
+          references(:organisation, type: :uuid, column: :id, on_delete: :nothing),
+          null: false
+        )
 
-      modify(:subscriber_id, references(:user, type: :uuid, column: :id, on_delete: :nothing))
+        modify(:subscriber_id, references(:user, type: :uuid, column: :id, on_delete: :nothing))
+      end
     end
   end
 end
